+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './advertising';
+export * from './analytics';
 export * from './audioSession';
 export * from './components';
 export * from './drm';
@@ -7,6 +8,8 @@ export * from './hooks';
 export * from './player';
 export * from './source';
 export * from './subtitleTrack';
+export * from './styleConfig';
+export * from './ui';
 export * from './offlineContentOptions';
 export * from './offlineContentManager';
 export * from './offlineContentManagerListener';